Change an array index variable to a size_t
Apparently using an unsigned int to access an array is a bad idea on
platforms where unsigned int does not match the size of a pointer,
because extra code is generated to properly enforce C's defined
unsigned integer overflow behaviour.

+++ src/mixer.c
@@ -415,7 +415,7 @@ CLOWNAUDIO_EXPORT ClownAudio_Sound* ClownAudio_Mixer_SoundCreate(ClownAudio_Mixe
 
 		void *resampled_decoders[2] = {NULL, NULL};
 
-		for (unsigned int i = 0; i < 2; ++i)
+		for (size_t i = 0; i < 2; ++i)
 		{
 			if (decoder_selectors[i] != NULL)
 			{
